+++ spec/fixtures/cookbooks/oneview_test/recipes/logical_interconnect_group_create_if_missing.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: oneview_test
-# Recipe:: firmware_bundle_add_invalid
+# Recipe:: logical_interconnect_group_create_if_missing
 #
 # (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
@@ -14,6 +14,10 @@
 # specific language governing permissions and limitations under the License.
 #
 
-oneview_firmware_bundle 'cp027376.exe' do
+oneview_logical_interconnect_group 'LogicalInterconnectGroup2' do
   client node['oneview_test']['client']
+  data(
+    enclosureType: 'C7000'
+  )
+  action :create_if_missing
 end