+++ spec/fixtures/cookbooks/oneview_test/recipes/rack_add_if_missing.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: oneview_test
-# Recipe:: firmware_bundle_add
+# Recipe:: rack_add_if_missing
 #
 # (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
@@ -14,7 +14,7 @@
 # specific language governing permissions and limitations under the License.
 #
 
-oneview_firmware_bundle 'upload cp027376.exe' do
+oneview_rack 'Rack1' do
   client node['oneview_test']['client']
-  file_path '/path/to/cp027376.exe'
+  action :add_if_missing
 end