+++ spec/fixtures/cookbooks/oneview_test/recipes/rack_remove_from_rack.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: oneview_test
-# Recipe:: firmware_bundle_add
+# Recipe:: rack_remove_from_rack
 #
 # (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
@@ -14,7 +14,11 @@
 # specific language governing permissions and limitations under the License.
 #
 
-oneview_firmware_bundle 'upload cp027376.exe' do
+oneview_rack 'Rack1' do
   client node['oneview_test']['client']
-  file_path '/path/to/cp027376.exe'
+  mount_options(
+    name: 'Encl1',
+    type: 'Enclosure'
+  )
+  action :remove_from_rack
 end