+++ spec/fixtures/cookbooks/oneview_test/recipes/server_hardware_add_if_missing.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: oneview_test
-# Recipe:: firmware_bundle_add
+# Recipe:: server_hardware_add_if_missing
 #
 # (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
@@ -14,7 +14,14 @@
 # specific language governing permissions and limitations under the License.
 #
 
-oneview_firmware_bundle 'upload cp027376.exe' do
+oneview_server_hardware 'ServerHardware1' do
+  data(
+    hostname: '172.18.6.4',
+    username: 'dcs',
+    password: 'dcs',
+    licensingIntent: 'OneViewStandard',
+    configurationState: 'Monitored'
+  )
   client node['oneview_test']['client']
-  file_path '/path/to/cp027376.exe'
+  action :add_if_missing
 end