+++ spec/unit/resources/firmware_bundle/add_spec.rb
@@ -7,7 +7,7 @@ describe 'oneview_test::firmware_bundle_add' do
   it 'adds it when it does not exist' do
     allow(OneviewSDK::FirmwareDriver).to receive(:find_by).and_return([])
     expect(OneviewSDK::FirmwareBundle).to receive(:add).and_return(OneviewSDK::FirmwareDriver.new(client, {}))
-    expect(real_chef_run).to add_oneview_firmware_bundle('cp027376.exe')
+    expect(real_chef_run).to add_oneview_firmware_bundle('upload cp027376.exe')
   end
 
   it 'does nothing when it already exists' do
@@ -17,15 +17,6 @@ describe 'oneview_test::firmware_bundle_add' do
       ]
     )
     expect(OneviewSDK::FirmwareBundle).to_not receive(:add)
-    expect(real_chef_run).to add_oneview_firmware_bundle('cp027376.exe')
-  end
-end
-
-describe 'oneview_test::firmware_bundle_add_invalid' do
-  let(:resource_name) { 'firmware_bundle' }
-  include_context 'chef context'
-
-  it 'fails if file_path is not set' do
-    expect { real_chef_run }.to raise_error(/Unspecified property: 'file_path'/)
+    expect(real_chef_run).to add_oneview_firmware_bundle('upload cp027376.exe')
   end
 end