+++ CHANGELOG.md
@@ -9,7 +9,10 @@
   - Added support to Logical enclosure `:reconfigure` and `:set_script`
   - Added support to Network set actions
   - Added support to Datacenter actions
+  - Added support to Rack actions
+  - Added support to Server hardware actions
   - Added support to Interconnect actions
+  - Added support to Logical switch group actions
   - Added support to Firmware bundle actions
 
 ### 0.1.1